airline-route-ranker/backend/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
email-validator>=2.0.0  # For EmailStr validation

# External API Communication
requests>=2.28.2
python-dotenv>=1.0.0

# Database
supabase>=1.0.3
psycopg2-binary>=2.9.6  # For direct PostgreSQL connections

# Data Processing
pandas>=2.0.0

# Utilities
python-dateutil>=2.8.2